Decouple registry and repository
