+++ mlserver/repository.py
@@ -1,90 +1,55 @@
-import asyncio
+import os
+import glob
 
 from typing import List
-from itertools import chain
 
-from .model import MLModel
-from .errors import ModelNotFound
-from .types import RepositoryIndexResponse
+from .settings import ModelParameters, ModelSettings
+
+DEFAULT_MODEL_SETTINGS_FILENAME = "model-settings.json"
 
 
-class SingleModelRepository:
+class ModelRepository:
     """
-    Repository for a single model with multiple versions.
+    Model repository, responsible for discovery of models which can be loaded
+    onto the model registry.
     """
 
-    def __init__(self, model: MLModel):
-        self._versions = {}
-        self._name = model.name
+    def __init__(self, root: str = None):
+        self._root = root
 
-        self._register(model)
+    def list(self) -> List[ModelSettings]:
+        all_model_settings = []
 
-    async def index(self) -> RepositoryIndexResponse:
+        # TODO: Use an async alternative for filesys ops
+        if self._root:
+            pattern = os.path.join(self._root, "**", DEFAULT_MODEL_SETTINGS_FILENAME)
+            matches = glob.glob(pattern, recursive=True)
+
+            for model_settings_path in matches:
+                model_settings = self._load_model_settings(model_settings_path)
+                all_model_settings.append(model_settings)
+
+        # If there were no matches, try to load model from environment
+        if not all_model_settings:
+            # return default
+            model_settings = ModelSettings()
+            model_settings.parameters = ModelParameters()
+            all_model_settings.append(model_settings)
+
+        return all_model_settings
+
+    def _load_model_settings(self, model_settings_path: str) -> ModelSettings:
+        model_settings = ModelSettings.parse_file(model_settings_path)
+
+        if not model_settings.parameters:
+            model_settings.parameters = ModelParameters()
+
+        if not model_settings.parameters.uri:
+            # If not specified, default to its own folder
+            default_model_uri = os.path.dirname(model_settings_path)
+            model_settings.parameters.uri = default_model_uri
+
+        return model_settings
+
+    def find(self, name: str) -> ModelSettings:
         pass
-
-    async def load(self, model: MLModel):
-        await model.load()
-        self._register(model)
-
-    async def get_model(self, version: str = None) -> MLModel:
-        if version:
-            if version not in self._versions:
-                raise ModelNotFound(self._name, version)
-
-            return self._versions[version]
-
-        return self._default
-
-    async def get_models(self) -> List[MLModel]:
-        models = self._versions.values()
-
-        # Add default if not versioned
-        if not self._default.version:
-            models.append(self._default)
-
-        return models
-
-    def _register(self, model: MLModel):
-        if model.version:
-            self._versions[model.version] = model
-
-        # TODO: Support version policies
-        self._default = model
-
-
-class MultiModelRepository:
-    """
-    Multiple model repository, where each model can have multiple versions.
-    """
-
-    def __init__(self):
-        self._models = {}
-
-    async def index(self) -> RepositoryIndexResponse:
-        pass
-
-    async def load(self, model: MLModel):
-        if model.name not in self._models:
-            self._models[model.name] = SingleModelRepository(model)
-
-        await self._models[model.name].load(model)
-
-    async def unload(self, name: str):
-        if name not in self._models:
-            raise ModelNotFound(name)
-
-        del self._models[name]
-
-    async def get_model(self, name: str, version: str = None) -> MLModel:
-        if name not in self._models:
-            raise ModelNotFound(name, version)
-
-        model = await self._models[name].get_model(version)
-        return model
-
-    async def get_models(self) -> List[MLModel]:
-        models_list = await asyncio.gather(
-            *[model.get_models() for model in self._models.values()]
-        )
-
-        return chain.from_iterable(models_list)
