+++ mlserver/server.py
@@ -5,7 +5,7 @@ from typing import List
 
 from .model import MLModel
 from .settings import Settings
-from .repository import MultiModelRepository
+from .registry import MultiModelRegistry
 from .handlers import DataPlane
 from .rest import RESTServer
 from .grpc import GRPCServer
@@ -15,16 +15,18 @@ HANDLED_SIGNALS = [signal.SIGINT, signal.SIGTERM]
 
 class MLServer:
     def __init__(self, settings: Settings):
-        self._model_repository = MultiModelRepository()
+        self._model_registry = MultiModelRegistry()
         self._settings = settings
         self._data_plane = DataPlane(
-            settings=self._settings, model_repository=self._model_repository
+            settings=self._settings, model_registry=self._model_registry
         )
 
     async def start(self, models: List[MLModel] = []):
         self._add_signal_handlers()
 
-        load_tasks = [self._model_repository.load(model) for model in models]
+        # TODO: Discover models from ModelRepository
+        # TODO: Add flag to disable autoload of models
+        load_tasks = [self._model_registry.load(model) for model in models]
         await asyncio.gather(*load_tasks)
 
         self._rest_server = RESTServer(self._settings, self._data_plane)
