+++ tests/test_registry.py
@@ -1,0 +1,31 @@
+import pytest
+
+from mlserver.errors import ModelNotFound
+
+
+@pytest.mark.parametrize(
+    "name, version",
+    [
+        ("sum-model", "v0"),
+        ("sum-model-2", "v0"),
+        ("sum-model", "v2"),
+        ("sum-model-2", None),
+    ],
+)
+async def test_get_model_not_found(model_registry, name, version):
+    with pytest.raises(ModelNotFound) as err:
+        await model_registry.get_model(name, version)
+
+        if version is not None:
+            assert err.message == f"Model {name} with version {version} not found"
+        else:
+            assert err.message == f"Model {name} not found"
+
+
+@pytest.mark.parametrize(
+    "name, version",
+    [("sum-model", "v1.2.3"), ("sum-model", None), ("sum-model", "")],
+)
+async def test_get_model(model_registry, sum_model, name, version):
+    found_model = await model_registry.get_model(name, version)
+    assert found_model == sum_model
